requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
mysqlclient==2.2.7
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
zope.interface==8.0.1
reportlab==4.2.5
